// apps/web/src/features/editor/types.ts
export interface FileTreeNode extends BackendFileNode {
  children?: Array<FileTreeNode>
}

export type FileNodeFormat = NonNullable<FileTreeNode["format"]>

export interface NodeProps extends NodeRendererProps<FileTreeNode> {
  onCreateFile: (
    parentId: string,
    index: number,
    isDir: boolean,
    format?: FileNodeFormat,
  ) => void
}